Corrected variables and added logger.error messages

+++ MongoDB/node.deleteMany.js
@@ -6,14 +6,20 @@ module.exports = async (connectorData, inputData) => {
     try {
         logger.trace(`MongoDB DeleteMany Node: Input - ${JSON.stringify(inputData)}`);
 
+        if (!inputData.collection) {
+            logger.error('MongoDB DeleteMany Node: Collection name is missing');
+            throw new Error('Collection name is missing');
+        }
+
         if (_.isEmpty(inputData.filter)) {
+            logger.error('MongoDB DeleteMany Node: Filter is empty');
             throw new Error('Filter is empty');
         }
 
         const result = await connectorData.db.collection(inputData.collection).deleteMany(inputData.filter);
 
         logger.trace(`MongoDB DeleteMany Node: Delete result - ${JSON.stringify(result)}`);
-        return { deleted_count: result.deletedCount };
+        return { deletedCount: result.deletedCount };
     } catch (error) {
         logger.error(`MongoDB DeleteMany Node: Error deleting documents from MongoDB: ${error}`);
         throw {
