+++ MongoDB/node.deleteOne.js
@@ -6,19 +6,25 @@ module.exports = async (connectorData, inputData) => {
     try {
         logger.trace(`MongoDB DeleteOne Node: Input - ${JSON.stringify(inputData)}`);
 
+        if (!inputData.collection) {
+            logger.error('MongoDB DeleteOne Node: Collection name is missing');
+            throw new Error('Collection name is missing');
+        }
+
         if (_.isEmpty(inputData.filter)) {
+            logger.error('MongoDB DeleteOne Node: Filter is empty');
             throw new Error('Filter is empty');
         }
 
         const result = await connectorData.db.collection(inputData.collection).deleteOne(inputData.filter);
 
         logger.trace(`MongoDB DeleteOne Node: Delete result - ${JSON.stringify(result)}`);
-        return { deleted_count: result.deletedCount };
+        return { deletedCount: result.deletedCount };
     } catch (error) {
-        logger.error(`MongoDB DeleteOne Node: Error deleting a document from MongoDB: ${error}`);
+        logger.error(`MongoDB DeleteOne Node: Error deleting document from MongoDB: ${error}`);
         throw {
             "code": "MONGODB_DELETE_ONE_ERROR",
-            "message": "Error deleting a document from MongoDB",
+            "message": "Error deleting document from MongoDB",
             "stackTrace": error
         };
     }
